Character level embeddings
Attempting to mimic embeddings at the character level.
Merging vocabularies from two corpora yielded weird results when training on a corpus that didn't include all characters in the vocabulary. I suppose it's because the unused inputs yielded dead connections between the input layer and the LSTM, and the unused characters are still sampled from when generating output.

Added fully connected layer just after the input. Had to resize the weight matrix (rank 2) so that it could be multiplied by the batch placeholder (rank 3).